# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchRevertRequest": {
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/BatchRevertOperation"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "operations"
        ],
        "title": "BatchRevertRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MergeHintRequest": {
        "properties": {
          "base_content": {
            "maxLength": 20000,
            "title": "Base Content",
            "type": "string"
          },
          "proposed_content": {
            "maxLength": 20000,
            "title": "Proposed Content",
            "type": "string"
          }
        },
        "required": [
          "base_content",
          "proposed_content"
        ],
        "title": "MergeHintRequest",
        "type": "object"
      },
      "MergeHintResponse": {
        "properties": {
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "has_conflict": {
            "title": "Has Conflict",
            "type": "boolean"
          },
          "merged_content": {
            "title": "Merged Content",
            "type": "string"
          },
          "strategy": {
            "enum": [
              "proposed",
              "current",
              "conflict_markers"
            ],
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "current_version",
          "merged_content",
          "has_conflict",
          "strategy"
        ],
        "title": "MergeHintResponse",
        "type": "object"
      },
      "NoteSectionResponse": {
        "properties": {
          "care_note_id": {
            "format": "uuid",
            "title": "Care Note Id",
            "type": "string"
          },
          "clinic_id": {
            "format": "uuid",
            "title": "Clinic Id",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_role": {
            "enum": [
              "staff",
              "clinician",
              "system"
            ],
            "title": "Owner Role",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "section_type": {
            "title": "Section Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "enum": [
              "internal",
              "patient_facing"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "patient_id",
          "care_note_id",
          "section_type",
          "owner_role",
          "created_by",
          "visibility",
          "content",
          "current_version",
          "created_at",
          "updated_at"
        ],
        "title": "NoteSectionResponse",
        "type": "object"
      },
...
      "RevertRequest": {
        "properties": {
          "change_reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "source_version": {
            "exclusiveMinimum": 0.0,
            "title": "Source Version",
            "type": "integer"
          }
        },
        "required": [
          "source_version",
          "expected_version"
        ],
        "title": "RevertRequest",
        "type": "object"
      },
      "RevisionComparisonResponse": {
        "properties": {
          "current_content": {
            "title": "Current Content",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "has_changes": {
            "title": "Has Changes",
            "type": "boolean"
          },
          "resource_id": {
            "format": "uuid",
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "enum": [
              "entry",
              "section"
            ],
            "title": "Resource Type",
            "type": "string"
          },
          "selected_content": {
            "title": "Selected Content",
            "type": "string"
          },
          "selected_version": {
            "title": "Selected Version",
            "type": "integer"
          },
          "unified_diff": {
            "title": "Unified Diff",
            "type": "string"
          },
          "word_diff": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Word Diff",
            "type": "array"
          }
        },
        "required": [
          "resource_type",
          "resource_id",
          "selected_version",
          "current_version",
          "selected_content",
          "current_content",
          "has_changes",
          "unified_diff"
        ],
        "title": "RevisionComparisonResponse",
        "type": "object"
      },
      "RevisionResponse": {
        "properties": {
          "change_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          },
          "changed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By"
          },
          "changed_by_role": {
            "enum": [
              "patient",
              "staff",
              "clinician",
              "system"
            ],
            "title": "Changed By Role",
            "type": "string"
          },
          "content_snapshot": {
            "title": "Content Snapshot",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "resource_id": {
            "format": "uuid",
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "enum": [
              "entry",
              "section"
            ],
            "title": "Resource Type",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "resource_type",
          "resource_id",
          "version_number",
          "content_snapshot",
          "changed_by",
          "changed_by_role",
          "change_reason",
          "created_at"
        ],
        "title": "RevisionResponse",
        "type": "object"
      },
...
      "TimelineEntryResponse": {
        "properties": {
          "author_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Id"
          },
          "author_role": {
            "enum": [
              "patient",
              "staff",
              "clinician",
              "system"
            ],
            "title": "Author Role",
            "type": "string"
          },
          "clinic_id": {
            "format": "uuid",
            "title": "Clinic Id",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "entry_type": {
            "title": "Entry Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceRecordResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_record_id": {
            "format": "uuid",
            "title": "Source Record Id",
            "type": "string"
          },
          "visibility": {
            "enum": [
              "internal",
              "patient_facing"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "patient_id",
          "author_id",
          "author_role",
          "entry_type",
          "visibility",
          "content",
          "source_record_id",
          "current_version",
          "occurred_at"
        ],
        "title": "TimelineEntryResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Server-enforced API for the Nightingale longitudinal Care Note.",
    "title": "Nightingale Care Note API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/entries/batch-revert": {
      "post": {
        "operationId": "batch_revert_entries_entries_batch_revert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRevertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimelineEntryResponse"
                  },
                  "title": "Response Batch Revert Entries Entries Batch Revert Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Batch Revert Entries",
        "tags": [
          "revisions"
        ]
      }
    },
...
    "/entries/{entry_id}/merge-hint": {
      "post": {
        "operationId": "entry_merge_hint_entries__entry_id__merge_hint_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeHintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeHintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Entry Merge Hint",
        "tags": [
          "revisions"
        ]
      }
    },
    "/entries/{entry_id}/revert": {
      "post": {
        "operationId": "revert_entry_entries__entry_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revert Entry",
        "tags": [
          "revisions"
        ]
      }
    },
    "/entries/{entry_id}/versions": {
      "get": {
        "operationId": "list_entry_versions_entries__entry_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RevisionResponse"
                  },
                  "title": "Response List Entry Versions Entries  Entry Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Entry Versions",
        "tags": [
          "revisions"
        ]
      }
    },
    "/entries/{entry_id}/versions/{version_number}/comparison": {
      "get": {
        "operationId": "compare_entry_version_entries__entry_id__versions__version_number__comparison_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_number",
            "required": true,
            "schema": {
              "title": "Version Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Compare Entry Version",
        "tags": [
          "revisions"
        ]
      }
    },
...
    "/sections/{section_id}/revert": {
      "post": {
        "operationId": "revert_section_sections__section_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Section Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteSectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revert Section",
        "tags": [
          "revisions"
        ]
      }
    },
    "/sections/{section_id}/versions": {
      "get": {
        "operationId": "list_section_versions_sections__section_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Section Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RevisionResponse"
                  },
                  "title": "Response List Section Versions Sections  Section Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Section Versions",
        "tags": [
          "revisions"
        ]
      }
    },
    "/sections/{section_id}/versions/{version_number}/comparison": {
      "get": {
        "operationId": "compare_section_version_sections__section_id__versions__version_number__comparison_get",
        "parameters": [
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Section Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_number",
            "required": true,
            "schema": {
              "title": "Version Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Compare Section Version",
        "tags": [
          "revisions"
        ]
      }
    },
...
  }
}